requirements.txt
streamlit
PyMuPDF
python-docx
spacy==3.7.2
sentence-transformers
langchain
langchain-google-genai
pandas